Move block types into 'types' module
