+++ bee-inx/src/block/mod.rs
@@ -1,68 +1,12 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod metadata;
+pub mod types;
 
-use bee_block as bee;
-use inx::proto;
+pub use self::types::*;
 
-pub use self::metadata::*;
-use crate::{return_err_if_none, Raw};
+use crate::client::Inx;
 
-/// The [`Block`] type.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct Block {
-    /// The [`BlockId`](bee::BlockId) of the block.
-    pub block_id: bee::BlockId,
-    /// The complete [`Block`](bee::Block) as raw bytes.
-    pub block: Raw<bee::Block>,
-}
-
-/// The [`BlockWithMetadata`] type.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct BlockWithMetadata {
-    /// The [`Metadata`](crate::BlockMetadata) of the block.
-    pub metadata: crate::BlockMetadata,
-    /// The complete [`Block`](bee::Block) as raw bytes.
-    pub block: Raw<bee::Block>,
-}
-
-impl TryFrom<proto::BlockWithMetadata> for BlockWithMetadata {
-    type Error = bee::InxError;
-
-    fn try_from(value: proto::BlockWithMetadata) -> Result<Self, Self::Error> {
-        Ok(BlockWithMetadata {
-            metadata: return_err_if_none!(value.metadata).try_into()?,
-            block: return_err_if_none!(value.block).data.into(),
-        })
-    }
-}
-
-impl From<BlockWithMetadata> for proto::BlockWithMetadata {
-    fn from(value: BlockWithMetadata) -> Self {
-        Self {
-            metadata: Some(value.metadata.into()),
-            block: Some(value.block.into()),
-        }
-    }
-}
-
-impl TryFrom<proto::Block> for Block {
-    type Error = bee::InxError;
-
-    fn try_from(value: proto::Block) -> Result<Self, Self::Error> {
-        Ok(Block {
-            block_id: return_err_if_none!(value.block_id).try_into()?,
-            block: return_err_if_none!(value.block).data.into(),
-        })
-    }
-}
-
-impl From<Block> for proto::Block {
-    fn from(value: Block) -> Self {
-        Self {
-            block_id: Some(value.block_id.into()),
-            block: Some(value.block.into()),
-        }
-    }
+impl Inx {
+    // TODO
 }
